Basic rating and setting functionality added.
Ratings now use the current date when selected from the main menu. The possible ratings now come from the database. If a rating and note exists for a given date the rating and note text views and edits will be filled accordingly. Next step is passing a date to the rating instead of having the rating activity to it by itself (required for selecting viewing calendar events/ratings).
Settings are also mostly implemented. The spinner is populated using the database. On change of the selection the text edits are filled accordingly. Any changes to the text or color are applied on press of the apply button. The rating should not be used at the moment, as the DatabaseHelper does not check between update and insertion. For this a seperate "Add rating" function should be added that is different from editing to make it clear that a new rating will be created. The alternative would be to handle it like ratings. If a said rating option does not exist it is added, else the entry is updated.
All past entries can also be deleted using the "Delete history button".
Calendar functionalities next.