Added a color picker and option to reset settings as well as other minor adjustments.
Colors can now be selected using a color picker. For the dialog a custom style has been created.
Settings can now be reset after changing.
The interface got some minor adjustments. For example the calendar now shows the month and year at the top as well as on the side. Unfortunately I have not found a way to hide the original month and year text.