Fixed color text field bug and settings reset error and some interface adjustments. Also lowered minimum SDK.
When the color text was populated with a non existent color and said color was applied somewhere the program would crash everytime it was supposed to be shown, e.g. the history and settings views. The colors cannot be updated before a correct color is chosen.
Before when resetting the settings all ratings would be deleted as well because the delete ratings query was wrong. Now has been updated so the original ratings between 1 and 5 remain.
Some adjustments to the interfaces, button and position adjustments mostly.
Minimum SDK now lowered to 19 to support Android 4.4 and up (98.1% of devices).