Fixed settings delete and update bugs, changed text color of rating input in darkmode.
When deleting all settings previously the program would crash because the entries were populated with non existent values. When creating or updating settings with no values the activity would crash because NullPointer and NumberExceptions weren't handled.
The text of the rating activity was white in dark mode on a white background. Now set to default #3C3C3C grey.